Fix test compile issues (EVM API modified)
