+++ params/protocol_params.go
@@ -233,5 +233,5 @@ var (
 
 	// RIP-7859 - Expose L1 origin information inside L2 execution environment
 	L1OriginContractAddress = common.HexToAddress("0x0000000000000000000000000000000000007859") // TBD address
-	L1OriginContractCode    = common.FromHex("608060405234801561000f575f5ffd5b50600436106100cd575f3560e01c8063776891fe1161008a578063c2a99ab411610064578063c2a99ab414610225578063d9f315fe14610243578063de8373e514610273578063dfe5743f14610291576100cd565b8063776891fe146101bb5780637f1bcccd146101d9578063856d71b914610209576100cd565b8063027d6dbb146100d15780633434735f1461010157806336b00a721461011f57806336b21b031461014f57806364f492d11461016d5780636c78d03c1461019d575b5f5ffd5b6100eb60048036038101906100e691906107d3565b6102af565b6040516100f89190610816565b60405180910390f35b6101096102c4565b604051610116919061086e565b60405180910390f35b610139600480360381019061013491906107d3565b6102dc565b6040516101469190610816565b60405180910390f35b6101576102f1565b6040516101649190610816565b60405180910390f35b610187600480360381019061018291906107d3565b61034a565b6040516101949190610816565b60405180910390f35b6101a561035f565b6040516101b29190610816565b60405180910390f35b6101c36103b8565b6040516101d09190610816565b60405180910390f35b6101f360048036038101906101ee91906107d3565b610410565b6040516102009190610816565b60405180910390f35b610223600480360381019061021e91906108b1565b610424565b005b61022d6105c6565b60405161023a9190610816565b60405180910390f35b61025d600480360381019061025891906107d3565b61061f565b60405161026a9190610816565b60405180910390f35b61027b610634565b6040516102889190610949565b60405180910390f35b61029961063d565b6040516102a69190610816565b60405180910390f35b5f6102b982610696565b600401549050919050565b73fffffffffffffffffffffffffffffffffffffffe81565b5f6102e682610696565b600101549050919050565b5f5f60015411610336576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161032d906109e2565b60405180910390fd5b610341600154610696565b60010154905090565b5f61035482610696565b600301549050919050565b5f5f600154116103a4576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161039b906109e2565b60405180910390fd5b6103af600154610696565b60030154905090565b5f5f600154116103fd576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016103f4906109e2565b60405180910390fd5b610408600154610696565b5f0154905090565b5f61041a82610696565b5f01549050919050565b73fffffffffffffffffffffffffffffffffffffffe73ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16146104a6576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161049d90610a70565b60405180910390fd5b5f86116104e8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016104df90610afe565b60405180910390fd5b5f612000876104f79190610b49565b90506040518060c00160405280878152602001868152602001858152602001848152602001838152602001888152505f5f8381526020019081526020015f205f820151815f01556020820151816001015560408201518160020155606082015181600301556080820151816004015560a0820151816005015590505060015487111561058557866001819055505b867f82e7f14988d91d60d054a4f1f35143fa0ae61e1721ccc4d10f20da0da37a507e876040516105b59190610816565b60405180910390a250505050505050565b5f5f6001541161060b576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610602906109e2565b60405180910390fd5b610616600154610696565b60040154905090565b5f61062982610696565b600201549050919050565b5f600154905090565b5f5f60015411610682576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610679906109e2565b60405180910390fd5b61068d600154610696565b60020154905090565b5f5f82116106d9576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016106d090610afe565b60405180910390fd5b60015482111561071e576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161071590610be9565b60405180910390fd5b5f6120008361072d9190610b49565b9050825f5f8381526020019081526020015f206005015414610784576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161077b90610c77565b60405180910390fd5b5f5f8281526020019081526020015f20915050919050565b5f5ffd5b5f819050919050565b6107b2816107a0565b81146107bc575f5ffd5b50565b5f813590506107cd816107a9565b92915050565b5f602082840312156107e8576107e761079c565b5b5f6107f5848285016107bf565b91505092915050565b5f819050919050565b610810816107fe565b82525050565b5f6020820190506108295f830184610807565b92915050565b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f6108588261082f565b9050919050565b6108688161084e565b82525050565b5f6020820190506108815f83018461085f565b92915050565b610890816107fe565b811461089a575f5ffd5b50565b5f813590506108ab81610887565b92915050565b5f5f5f5f5f5f60c087890312156108cb576108ca61079c565b5b5f6108d889828a016107bf565b96505060206108e989828a0161089d565b95505060406108fa89828a0161089d565b945050606061090b89828a0161089d565b935050608061091c89828a0161089d565b92505060a061092d89828a0161089d565b9150509295509295509295565b610943816107a0565b82525050565b5f60208201905061095c5f83018461093a565b92915050565b5f82825260208201905092915050565b7f4c314f726967696e536f75726365496d706c3a206e6f204c3120626c6f636b735f8201527f20617661696c61626c6500000000000000000000000000000000000000000000602082015250565b5f6109cc602a83610962565b91506109d782610972565b604082019050919050565b5f6020820190508181035f8301526109f9816109c0565b9050919050565b7f4c314f726967696e536f75726365496d706c3a2063616c6c6572206973206e6f5f8201527f74207468652073797374656d2061646472657373000000000000000000000000602082015250565b5f610a5a603483610962565b9150610a6582610a00565b604082019050919050565b5f6020820190508181035f830152610a8781610a4e565b9050919050565b7f4c314f726967696e536f75726365496d706c3a206865696768742063616e6e6f5f8201527f74206265207a65726f0000000000000000000000000000000000000000000000602082015250565b5f610ae8602983610962565b9150610af382610a8e565b604082019050919050565b5f6020820190508181035f830152610b1581610adc565b9050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601260045260245ffd5b5f610b53826107a0565b9150610b5e836107a0565b925082610b6e57610b6d610b1c565b5b828206905092915050565b7f4c314f726967696e536f75726365496d706c3a20626c6f636b206865696768745f8201527f20746f6f20686967680000000000000000000000000000000000000000000000602082015250565b5f610bd3602983610962565b9150610bde82610b79565b604082019050919050565b5f6020820190508181035f830152610c0081610bc7565b9050919050565b7f4c314f726967696e536f75726365496d706c3a20626c6f636b2064617461206e5f8201527f6f7420666f756e64206f72206f7665727772697474656e000000000000000000602082015250565b5f610c61603783610962565b9150610c6c82610c07565b604082019050919050565b5f6020820190508181035f830152610c8e81610c55565b905091905056fea2646970667358221220617ab035478163881a9c089ebdd70ef3e52fb621569e019cad9f3bad6ccee6f164736f6c634300081d0033")
+	L1OriginContractCode    = common.FromHex("608060405234801561000f575f5ffd5b50600436106100cd575f3560e01c8063776891fe1161008a578063c2a99ab411610064578063c2a99ab414610225578063d9f315fe14610243578063de8373e514610273578063dfe5743f14610291576100cd565b8063776891fe146101bb5780637f1bcccd146101d9578063856d71b914610209576100cd565b8063027d6dbb146100d15780633434735f1461010157806336b00a721461011f57806336b21b031461014f57806364f492d11461016d5780636c78d03c1461019d575b5f5ffd5b6100eb60048036038101906100e69190610927565b6102af565b6040516100f8919061096a565b60405180910390f35b6101096102c4565b60405161011691906109c2565b60405180910390f35b61013960048036038101906101349190610927565b6102dc565b604051610146919061096a565b60405180910390f35b6101576102f1565b604051610164919061096a565b60405180910390f35b61018760048036038101906101829190610927565b61034a565b604051610194919061096a565b60405180910390f35b6101a561035f565b6040516101b2919061096a565b60405180910390f35b6101c36103b8565b6040516101d0919061096a565b60405180910390f35b6101f360048036038101906101ee9190610927565b610410565b604051610200919061096a565b60405180910390f35b610223600480360381019061021e9190610a05565b610424565b005b61022d61071a565b60405161023a919061096a565b60405180910390f35b61025d60048036038101906102589190610927565b610773565b60405161026a919061096a565b60405180910390f35b61027b610788565b6040516102889190610a9d565b60405180910390f35b610299610791565b6040516102a6919061096a565b60405180910390f35b5f6102b9826107ea565b600401549050919050565b73fffffffffffffffffffffffffffffffffffffffe81565b5f6102e6826107ea565b600101549050919050565b5f5f60015411610336576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161032d90610b36565b60405180910390fd5b6103416001546107ea565b60010154905090565b5f610354826107ea565b600301549050919050565b5f5f600154116103a4576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161039b90610b36565b60405180910390fd5b6103af6001546107ea565b60030154905090565b5f5f600154116103fd576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016103f490610b36565b60405180910390fd5b6104086001546107ea565b5f0154905090565b5f61041a826107ea565b5f01549050919050565b73fffffffffffffffffffffffffffffffffffffffe73ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16146104a6576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161049d90610bc4565b60405180910390fd5b5f86116104e8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016104df90610c52565b60405180910390fd5b5f5f1b851161052c576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161052390610ce0565b60405180910390fd5b5f5f1b8411610570576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161056790610d6e565b60405180910390fd5b5f5f1b83116105b4576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016105ab90610dfc565b60405180910390fd5b5f5f1b82116105f8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016105ef90610e8a565b60405180910390fd5b5f5f1b811161063c576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161063390610f18565b60405180910390fd5b5f6120008761064b9190610f63565b90506040518060c00160405280878152602001868152602001858152602001848152602001838152602001888152505f5f8381526020019081526020015f205f820151815f01556020820151816001015560408201518160020155606082015181600301556080820151816004015560a082015181600501559050506001548711156106d957866001819055505b867f82e7f14988d91d60d054a4f1f35143fa0ae61e1721ccc4d10f20da0da37a507e87604051610709919061096a565b60405180910390a250505050505050565b5f5f6001541161075f576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161075690610b36565b60405180910390fd5b61076a6001546107ea565b60040154905090565b5f61077d826107ea565b600201549050919050565b5f600154905090565b5f5f600154116107d6576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016107cd90610b36565b60405180910390fd5b6107e16001546107ea565b60020154905090565b5f5f821161082d576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161082490610c52565b60405180910390fd5b600154821115610872576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161086990611003565b60405180910390fd5b5f612000836108819190610f63565b9050825f5f8381526020019081526020015f2060050154146108d8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016108cf90611091565b60405180910390fd5b5f5f8281526020019081526020015f20915050919050565b5f5ffd5b5f819050919050565b610906816108f4565b8114610910575f5ffd5b50565b5f81359050610921816108fd565b92915050565b5f6020828403121561093c5761093b6108f0565b5b5f61094984828501610913565b91505092915050565b5f819050919050565b61096481610952565b82525050565b5f60208201905061097d5f83018461095b565b92915050565b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f6109ac82610983565b9050919050565b6109bc816109a2565b82525050565b5f6020820190506109d55f8301846109b3565b92915050565b6109e481610952565b81146109ee575f5ffd5b50565b5f813590506109ff816109db565b92915050565b5f5f5f5f5f5f60c08789031215610a1f57610a1e6108f0565b5b5f610a2c89828a01610913565b9650506020610a3d89828a016109f1565b9550506040610a4e89828a016109f1565b9450506060610a5f89828a016109f1565b9350506080610a7089828a016109f1565b92505060a0610a8189828a016109f1565b9150509295509295509295565b610a97816108f4565b82525050565b5f602082019050610ab05f830184610a8e565b92915050565b5f82825260208201905092915050565b7f4c314f726967696e536f75726365496d706c3a206e6f204c3120626c6f636b735f8201527f20617661696c61626c6500000000000000000000000000000000000000000000602082015250565b5f610b20602a83610ab6565b9150610b2b82610ac6565b604082019050919050565b5f6020820190508181035f830152610b4d81610b14565b9050919050565b7f4c314f726967696e536f75726365496d706c3a2063616c6c6572206973206e6f5f8201527f74207468652073797374656d2061646472657373000000000000000000000000602082015250565b5f610bae603483610ab6565b9150610bb982610b54565b604082019050919050565b5f6020820190508181035f830152610bdb81610ba2565b9050919050565b7f4c314f726967696e536f75726365496d706c3a206865696768742063616e6e6f5f8201527f74206265207a65726f0000000000000000000000000000000000000000000000602082015250565b5f610c3c602983610ab6565b9150610c4782610be2565b604082019050919050565b5f6020820190508181035f830152610c6981610c30565b9050919050565b7f4c314f726967696e536f75726365496d706c3a20626c6f636b486173682063615f8201527f6e6e6f74206265207a65726f0000000000000000000000000000000000000000602082015250565b5f610cca602c83610ab6565b9150610cd582610c70565b604082019050919050565b5f6020820190508181035f830152610cf781610cbe565b9050919050565b7f4c314f726967696e536f75726365496d706c3a20706172656e74426561636f6e5f8201527f526f6f742063616e6e6f74206265207a65726f00000000000000000000000000602082015250565b5f610d58603383610ab6565b9150610d6382610cfe565b604082019050919050565b5f6020820190508181035f830152610d8581610d4c565b9050919050565b7f4c314f726967696e536f75726365496d706c3a207374617465526f6f742063615f8201527f6e6e6f74206265207a65726f0000000000000000000000000000000000000000602082015250565b5f610de6602c83610ab6565b9150610df182610d8c565b604082019050919050565b5f6020820190508181035f830152610e1381610dda565b9050919050565b7f4c314f726967696e536f75726365496d706c3a2072656365697074526f6f74205f8201527f63616e6e6f74206265207a65726f000000000000000000000000000000000000602082015250565b5f610e74602e83610ab6565b9150610e7f82610e1a565b604082019050919050565b5f6020820190508181035f830152610ea181610e68565b9050919050565b7f4c314f726967696e536f75726365496d706c3a207472616e73616374696f6e525f8201527f6f6f742063616e6e6f74206265207a65726f0000000000000000000000000000602082015250565b5f610f02603283610ab6565b9150610f0d82610ea8565b604082019050919050565b5f6020820190508181035f830152610f2f81610ef6565b9050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601260045260245ffd5b5f610f6d826108f4565b9150610f78836108f4565b925082610f8857610f87610f36565b5b828206905092915050565b7f4c314f726967696e536f75726365496d706c3a20626c6f636b206865696768745f8201527f20746f6f20686967680000000000000000000000000000000000000000000000602082015250565b5f610fed602983610ab6565b9150610ff882610f93565b604082019050919050565b5f6020820190508181035f83015261101a81610fe1565b9050919050565b7f4c314f726967696e536f75726365496d706c3a20626c6f636b2064617461206e5f8201527f6f7420666f756e64206f72206f7665727772697474656e000000000000000000602082015250565b5f61107b603783610ab6565b915061108682611021565b604082019050919050565b5f6020820190508181035f8301526110a88161106f565b905091905056fea264697066735822122097d117e54ba752323cec0ba2bd8328ff99cd465cf2aac260ec8b69f4c43e32b764736f6c634300081d0033")
 )
